Linux build refactor

+++ eob2/avatar.h
@@ -3,6 +3,6 @@
 enum gendern : unsigned char;
 
 void avatar_read(const char* url);
-size_t get_avatars(unsigned char* result, char race, gendern gender, char cls);
+unsigned get_avatars(unsigned char* result, char race, gendern gender, char cls);
 unsigned char generate_avatar(char race, gendern gender, char cls);
-unsigned char generate_avatar(char race, gendern gender, char cls, unsigned char* exclude, size_t exclude_size);
+unsigned char generate_avatar(char race, gendern gender, char cls, unsigned char* exclude, unsigned long exclude_size);
